@@ -1,5 +1,5 @@
-require "middleman-core"
-require "middleman-robots/version"
+require 'middleman-core'
+require 'middleman-robots/version'
 
 ::Middleman::Extensions.register(:robots) do
   require 'middleman-robots/extension'
