@@ -1,6 +1,6 @@
 module Middleman
   # Middleman robots module
   module Robots
-    VERSION = '1.3.2'
+    VERSION = '1.3.2'.freeze
   end
 end
