@@ -1,25 +1,24 @@
 require 'minitest/autorun'
 require_relative '../lib/middleman-robots/group.rb'
 
-class Test_Group < MiniTest::Test
-
+class TestGroup < MiniTest::Test
   def test_initialize
     rule = {
       user_agent: 'GoogleBot',
-      disallow: %w(tmp/* /someting/dir/disallow.html),
-      allow: %w(allow/* /someting/dir/allow.html),
+      disallow: %w[tmp/* /someting/dir/disallow.html],
+      allow: %w[allow/* /someting/dir/allow.html]
     }
     group = Middleman::Robots::Group.new(rule)
 
     assert_equal rule[:user_agent], group.user_agent
-    assert_equal %w(/tmp/* /someting/dir/disallow.html), group.disallow
-    assert_equal %w(/allow/* /someting/dir/allow.html), group.allow
+    assert_equal %w[/tmp/* /someting/dir/disallow.html], group.disallow
+    assert_equal %w[/allow/* /someting/dir/allow.html], group.allow
   end
 
   def test_initialize_without_user_agent
     rule = {
-      disallow: %w(/tmp/*),
-      allow: %w(/allow/*),
+      disallow: %w[/tmp/*],
+      allow: %w[/allow/*]
     }
     group = Middleman::Robots::Group.new(rule)
 
@@ -31,7 +30,7 @@ class Test_Group < MiniTest::Test
   def test_initialize_without_disallow
     rule = {
       user_agent: 'GoogleBot',
-      allow: %w(/allow/* /someting/dir/allow.html),
+      allow: %w[/allow/* /someting/dir/allow.html]
     }
     group = Middleman::Robots::Group.new(rule)
 
@@ -43,7 +42,7 @@ class Test_Group < MiniTest::Test
   def test_initialize_without_allow
     rule = {
       user_agent: 'GoogleBot',
-      disallow: %w(/tmp/* /someting/dir/disallow.html),
+      disallow: %w[/tmp/* /someting/dir/disallow.html]
     }
     group = Middleman::Robots::Group.new(rule)
 
@@ -55,34 +54,34 @@ class Test_Group < MiniTest::Test
   def test_text
     rule = {
       user_agent: 'GoogleBot',
-      disallow: %w(tmp/* /someting/dir/disallow.html),
-      allow: %w(allow/* /someting/dir/allow.html),
+      disallow: %w[tmp/* /someting/dir/disallow.html],
+      allow: %w[allow/* /someting/dir/allow.html]
     }
     group = Middleman::Robots::Group.new(rule)
 
-    expected = <<-END
-User-Agent: GoogleBot
-Disallow: /tmp/*
-Disallow: /someting/dir/disallow.html
-Allow: /allow/*
-Allow: /someting/dir/allow.html
-    END
+    expected = <<~ROBOTS
+      User-Agent: GoogleBot
+      Disallow: /tmp/*
+      Disallow: /someting/dir/disallow.html
+      Allow: /allow/*
+      Allow: /someting/dir/allow.html
+    ROBOTS
 
     assert_equal expected, group.text
   end
 
   def test_text_without_user_agent
     rule = {
-      disallow: %w(/tmp/*),
-      allow: %w(/allow/*),
+      disallow: %w[/tmp/*],
+      allow: %w[/allow/*]
     }
     group = Middleman::Robots::Group.new(rule)
 
-    expected = <<-END
-User-Agent: *
-Disallow: /tmp/*
-Allow: /allow/*
-    END
+    expected = <<~ROBOTS
+      User-Agent: *
+      Disallow: /tmp/*
+      Allow: /allow/*
+    ROBOTS
 
     assert_equal expected, group.text
   end
@@ -90,15 +89,15 @@ Allow: /allow/*
   def test_text_without_disallow
     rule = {
       user_agent: 'GoogleBot',
-      allow: %w(/allow/* /someting/dir/allow.html),
+      allow: %w[/allow/* /someting/dir/allow.html]
     }
     group = Middleman::Robots::Group.new(rule)
 
-    expected = <<-END
-User-Agent: GoogleBot
-Allow: /allow/*
-Allow: /someting/dir/allow.html
-    END
+    expected = <<~ROBOTS
+      User-Agent: GoogleBot
+      Allow: /allow/*
+      Allow: /someting/dir/allow.html
+    ROBOTS
 
     assert_equal expected, group.text
   end
@@ -106,15 +105,15 @@ Allow: /someting/dir/allow.html
   def test_text_without_allow
     rule = {
       user_agent: 'GoogleBot',
-      disallow: %w(/tmp/* /someting/dir/disallow.html),
+      disallow: %w[/tmp/* /someting/dir/disallow.html]
     }
     group = Middleman::Robots::Group.new(rule)
 
-    expected = <<-END
-User-Agent: GoogleBot
-Disallow: /tmp/*
-Disallow: /someting/dir/disallow.html
-    END
+    expected = <<~ROBOTS
+      User-Agent: GoogleBot
+      Disallow: /tmp/*
+      Disallow: /someting/dir/disallow.html
+    ROBOTS
 
     assert_equal expected, group.text
   end
